requirements.txt
Django==2.0.5
django-extensions==2.0.7
django-extra-fields==1.0.0
djangorestframework==3.8.2
Pillow==5.1.0
pytz==2018.4
six==1.11.0

